requirements.txt
# Whisper dependencies
openai-whisper>=20231117
openai>=1.3.0
torch>=2.0.0
torchaudio>=2.0.0

# Audio processing
pydub>=0.25.1

# Existing dependencies
httpx>=0.24.0
python-docx>=0.8.11
python-dotenv>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
ruff>=0.1.0